103ms initial connection, 34 of waiting for response, 21ms of the real c++ logic for setting pixels colors... even if is only 130ms. In average I see that it send 2.x images each second, so each 420ms.
TODO: In the next commits, maybe I will not use http and return to websockets to solve the initial connection problem.
In the past I decided to not use websockets because of chunk division, but I learnt this weekend the importance of index, len, total params so maybe I can use those and solve the chunking problem.

another bug that I have now is that when I open double tabs you see that it have problems, it don't share the same state between them and so on. So I think that websocket can solve also that problem